// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type * as auth from "../auth.js";
import type * as authActions from "../authActions.js";
import type * as crons from "../crons.js";
import type * as notificationTypes from "../notificationTypes.js";
import type * as notifications from "../notifications.js";
import type * as notificationsNew from "../notificationsNew.js";
import type * as pushNotifications from "../pushNotifications.js";
import type * as rateLimit from "../rateLimit.js";
import type * as restaurantManagement from "../restaurantManagement.js";
import type * as restaurants from "../restaurants.js";
import type * as users from "../users.js";

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  auth: typeof auth;
  authActions: typeof authActions;
  crons: typeof crons;
  notificationTypes: typeof notificationTypes;
  notifications: typeof notifications;
  notificationsNew: typeof notificationsNew;
  pushNotifications: typeof pushNotifications;
  rateLimit: typeof rateLimit;
  restaurantManagement: typeof restaurantManagement;
  restaurants: typeof restaurants;
  users: typeof users;
}>;
declare const fullApiWithMounts: typeof fullApi;

export declare const api: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApiWithMounts,
  FunctionReference<any, "internal">
>;

export declare const components: {};